Init dynamic styles
