+++ libs/core/src/dynamic-styles.ts
@@ -1,0 +1,1 @@
+export * from './utils/dynamic-styles/dynamic-styles'
